feat: add achievement api methods
